Renamed files and added more code
