+++ console.py
@@ -20,7 +20,7 @@ class Console:
         print("%s", string)
         
     def getCoordinate(self, string):
-        """ Get coordinate from user or computer and validate it"""
+        """ Get coordinate from user or computer and validate it """
         """ string -> integer """
         
         running = True;
@@ -51,6 +51,10 @@ class Console:
         """ Pauses game """
         """ int -> void """
         time.sleep(int)
+        
+    def playAgain(self):
+        """ Checks if player wants to play again """
+        # TODO
         
     def showFleetStatus(self, board):
         """ Displays the number of ships still in play """
